switch auth backend to CAS + SAML
the DMT prod and staging URLs have been registered with our CAS service
with CUIT, so this should be good to go now.

The big caveat here is that with CAS, now to develop on the DMT locally,
if you need to login/logout, you will need to have your development
machine registered, and you'll need to run a local https proxy.

It's a pain, but it's something we have to get used to for working with CAS
apps.