manage command for moving attachments to S3
This is a one-time operation, which is why some stuff is hard-coded and doesn't come
with tests.

All it's doing is going through every attachment in the database that the PMT has
stored locally and shipping it up to the configured S3 bucket, using the same
path/filename scheme that we've established elsewhere. Then it updates the db
entry with the new S3 URL.

Once this goes out to production, I will rsync the uploaded attachments from hutz
(where the PMT lives) to the /mnt/nfs/applications/dmt/attachments/ directory,
then run './manage.py move_attachments_to_s3 --settings...' and that should
be that.

Once everything has been moved up and verified, I'll delete this command
since there will no longer be any use for it.