resolved items report
OK, a pet peeve of mine is when items sit in the RESOLVED state forever.
I don't understand why, but some people like to assign lots of tasks
to other people and then when they get resolved, they take their
sweet time to verify them.

This is a very simple report to name and shame.