Fix comment wrapping
This was an issue with chrome, but not firefox. Long comments wouldn't
get wrapped properly -- they were trailing off the screen. I couldn't
figure out a way to fix this in the <td> without using
"table-layout: fixed;" on the table, which caused other problems. I'm
using the bootstrap grid system for this now which is working better.