:green_heart: need to update factories/tests as well
yeah, didn't notice that a lot of the tests were written to instantiate
User objects directly instead of going through a factory. That needed to
be changed to use a factory that's set up to handle the OneToOne stuff
properly.

One view, for creating new groups also had to change a bit. With the OneToOne
django profile approach, it expects Users to be created first, before
a UserProfile. So, creating a Group (which is still a hacked up UserProfile)
has to work that way as well.