add nullable foreign key fields pointing to User
first step towards getting things working more like a regular django app
with the User as the central object and UserProfile as extra stuff attached to it.

After this has gone through, the next step will be to make every add/update that
touches a UserProfile foreign key also do the correct thing with this User key.

Then a data migration to backfill all the existing ones.

Then the fields can be set to not-null.

Then we can start using the User links instead of UserProfile links.