data migration to populate foreign key fields pointing to User
Everyone working on the DMT should take note of this one.
The data migration has to update basically every row in some of the big tables
with 100,000+ rows. This means it's very slow. Running it locally on a copy
of the production database took over half an hour.

This means that the rolf pushes will probably break/timeout so they
will have to be manually pushed through.

It also means that when you run 'make migrate' on your development instance,
it could take a very long time. Shouldn't be a big deal, but don't do that
if you were planning on working on it in the next few minutes or something. Run
the migration before you step out for lunch, or in the background while you work
on a different project.