:racehorse: optimize getting list of milestones with unverified items
The  seems to make Postgres not be able to figure out
how to use the index on that column, forcing a sequential scan over items every
time. So instead of doing one of those queries for each milestone in the project
we do it once for the whole project, make a set of the milestones, then
use that to filter on the python side.

On my test page, it shaves about 400ms off page load time. (that's with
20 milestones. more milestones in a project would mean more performance
improvement).