:racehorse: improve homepage load time
The recent additions of owned items and listing the owner on each
item in the list of outstanding items on the homepage were causing
a lot of extra queries and slowing things down if you have a lot
of items on the page. Added some select_related()s to pull that
info back into the top level queries and avoid the N+1 problems
that had crept in. On my homepage in testing, it reduced it from 250 queries
down to 34.