force it to wait until page is loaded before touching the db
The Django docs say:

When using an in-memory SQLite database to run the tests, the same database connection will be shared by two threads in parallel: the thread in which the live server is run and the thread in which the test case is run. It’s important to prevent simultaneous database queries via this shared connection by the two threads, as that may sometimes randomly cause the tests to fail. So you need to ensure that the two threads don’t access the database at the same time. In particular, this means that in some cases (for example, just after clicking a link or submitting a form), you might need to check that a response is received by Selenium and that the next page is loaded before proceeding with further test execution. Do this, for example, by making Selenium wait until the <body> HTML tag is found in the response

https://docs.djangoproject.com/en/1.7/topics/testing/tools/\#liveservertestcase

In my limited testing, this seems to clear out the 'OperationalError: cannot commit - no transaction is active'
problem that happens when create_pre_authenticated_session() races with the backend