Use postgresql when running tests locally
This is just an idea. We have a bunch of tests in dmt that
require postgresql, because of how it's written. I think that
they should run by default when running `make test`, since any
developer machine would have postgresql installed, since it's
required to run `make runserver`.

This changes the settings to use postgresql on `make test`, but
use sqlite on `make jenkins`.