Sort search results in search autocomplete - PMT #99496
When searching, for example, for "project", the projects
that start with the word "project" will name now appear
at the beginning of the search results.

I had to override a method in DRF's SearchFilter in order to
allow us to search on terms with spaces, (e.g. try searching
for "project m" to find the "Project Management Tool" project).
I might make a pull request upstream if I find no existing way to
configure this.