Add project report - pmt #99285
This commit adds postgresql-specific method - users_active_between()
to the Project model, and adds the result of that method to the
ProjectDetail context. This means that any test that accesses the
project detail page now needs a @skipUnless(postgresql) decorator.

I think this is okay, but if anyone disagrees I can look into
one of these alternatives:
- Implement users_active_between() in Django's ORM syntax. I couldn't
  think of a way to do this without making multiple queries, but maybe
  I'm missing something.
- Load the results of users_active_between() dynamically in JavaScript.

--

This commit also includes an update to bootstrap-datepicker, and a
tweak to how we're using jquery in requirejs.

I'm configuring the jQuery we're using with requirejs to not export
the global $ and jQuery variables, per these instructions here:
http://requirejs.org/docs/jquery.html#noconflictmap

Because I'm including bootstrap from outside of requirejs, bootstrap
was attaching itself to an inaccessible instance of jQuery.

The setup is now like this:
- A private jQuery, only accessible as an AMD module within requirejs.
- A global jQuery, which bootstrap attaches itself to

This necessitates wrapping the bootstrap-datepicker and typeahead
libraries in a define() call until these issues are resolved:
  https://github.com/twitter/typeahead.js/issues/922
  https://github.com/eternicode/bootstrap-datepicker/pull/1196