fix __init__() got an unexpected keyword argument 'template_name' error
The problem is that the `RangeOffsetMixin` has an `__init__()`, which overrides the
`__init__()` of the generic Django CBVs that lets you specify additional
parameters when you instantiate it in `urls.py`. That's what we do for
the project board view, since it uses the same data as the regular
project detail view, we just re-use that view and specify a different
`template_name` in `urls.py`.

Removed the `__init__()` from `RangeOffsetMixin` and just explicitly call
`calc_interval()` instead.'