remove per-milestone hours/estimates graphs
Every once in a while these are interesting and useful, but I'm
feeling like on the whole, the little stripchart graphs of
hours/estimated for every milestone just isn't a useful enough
feature to keep around. LITO's graphite server is slow and the result
is that they slow down loading of whole pages.