bugfix on Project.all_personnel_in_project(). PMT #98702
results were missing users in groups in the project. stupid 1 character typo.

Writing a test for the bug, I changed the default on the UserProfileFactory
to status='active', which makes a lot more sense.

That revealed a false positive on the test_views.GroupTest.test_remove_user_from_group()
test. It was checking for the existence of the username on the page, but even when
a user is removed, their username still appears in the 'Add User To Group' pulldown.

Switched it to check for the user's URL instead, indicating that they are actually
listed in the group, not just in a pulldown.