Add comment edit functionality - PMT #100766
This includes a new field on the Comment model: `comment_src`

I think it's more flexible if we store both the comment's
markdown source as well as its rendered HTML:
  http://stackoverflow.com/a/10456928/173630

Currently we're only storing the HTML.

I've included a migration that populates the new comment_src
field via html2text. Another way we could do this is on an
as-needed basis, in the view code, but I thought the all-at-once
migration was simpler. It took less than a minute to run locally,
once I narrowed the queried comments down to actual comments and
not all the events like "bug added".