Fix time calculation bug in PrevNextWeekMixin - PMT #101552
In the case where no URL param was provided, the current time
was having an effect on the reported trackers on the weekly report
mixin. To fix the issue, I'm calling calc_weeks() in this situation
and clearing out the time value, in order to calculate the beginning
of the week accurately.