Use the primary group for reporting - PMT #95814
Now we have primary groups and what I'm calling
"Other" groups, but we could change this to "Secondary"
if that's clearer.

The group detail page displays the primary members and the
other members, and allows you to add other members to the
selected group. Each user's primary group can be edited on
their profile page.

The staff report now reports on the user's "primary" group, instead
of an arbitrary "other" group.