Convert IntervalField to Django's builtin DurationField
This doesn't actually change the field type in postgresql
(it's still `interval`). But because this Django package
won't be maintained forever it would be good to switch
to something built into Django (as of Django 1.8).

I left django-pgsql-interval-field in requirements.txt for now
because the initial migration depends on it. Maybe we can get
rid of it by updating the initial migration to use DurationField?
That seems a little crazy so I just left it alone, but maybe
it's fine.