add pmt favicon
I'm not sure if we want to stick with the CCNMTL 'C' as
the favicon, but if we don't I thought we could have the honeycomb
icon as the favicon.

I made this with http://www.favicon.cc/, and I think it's an alright
rendition of the logo, but it could definitely be improved if we have
vector source files around.. maybe Zarina could help?

Other favicon generators, like http://realfavicongenerator.net/
require a larger source image, but give you larger favicons that
work well on tablets and phones.