Add search filtering for 'my items' table - PMT #100751
This adds a search box for the dashboard items in PMT next
to the show/hide columns dropdown.

In order for this to display correctly I needed to remove the
negative-margin hack on the show/hide columns dropdown, and now
the "weekly hours logged" graph is about this button and the search
input. I've also reduced some of the margins to make it look good.

Let me know if you don't like any of these changes, @zmustapha,
these are just some ideas. I really just want to be able to limit
the main "My Items" table by project, and this search box provided
by DataTables seems to work well.