Replace django-pgsql-interval-field with DurationField- PMT #103603
Replacing IntervalField with Django's built-in DurationField.

The only user-visible change here is that the item editing form shows a
different form field for "estimated time": a text input that displays
and accepts simpleduration-formatted strings.  I chose to use
simpleduration as the input format for this field, instead of making
a custom duration widget.
http://www.columbia.edu/~njn2118/journal/2015/10/30.html

As noted in the PMT item, my first attempt at this was incomplete
(it didn't contain a usable form field), and it also got reverted,
maybe due to migration issues? The migration seems to work on my
setup.