DST fix for reports.mixins.RangeOffsetMixin
the calc_interval() method doesn't return exactly what is expected when it straddles
a DST change. That's fine, really, but the tests break when that's the case.

To fix it, I made the tests use a set date that doesn't have that problem and
extended RangeOffsetMixin so the  value could be explicitly set (this
makes it easier to test with arbitrary dates in case we want to test the DST
behavior differently later).