link status updates to project page
continuing on PMT #104015

instead of linking to the edit form for a status update, the feed now
links to an appropriate id for the status update on the project page.

note that we still don't have js rigged up to switch to the updates tab
and highlight the linked item yet, but it's better than nothing.