unify Makefile and package.json functionality
Standardizing on using package.json to specify the JS libraries to be
installed so Makefile targets can just install everything with one 'npm
install' instead of scattering stuff around.

I also refactored the build dependencies a bit and added a sentinal
setup so more of this should be automatic (just update package.json and
everything else should happen automatically when needed).

The other thing worth noting is that instead of 'make js' you now would
do 'make media/main-built.js'. That's not as nice to type, but it's a
real file as a target so make can be smarter about whether or not to
rebuild it, which will be better for further deployment automation.