Fix daterange display when window < 1024px wide
When looking at the staff report in a window <1024px wide, the "Export
as" row was interfering with the daterange form's "Submit" button - it
was still visible, but most of it was clickable. I've changed around
a bootstrap grid setting in the daterange form and taken the export area
out of a `class="row"`, and the problem is fixed.