Make ActualTime.uuid non-null - PMT #105546
I'm really expecting these IDs to be unique, but currently, setting this
field as the primary key, or unique=True fails with duplicate errors.

For example, in my local database there's 4 ActualTimes with the uuid
`cf18fc3e-ea03-467c-aada-c9096160567d`.

This is just something to note, not a major problem, since it's not
often that a user will need to delete an ActualTime. Still, it's
unnerving since the chances of a UUID collision is incredibly small:
https://en.wikipedia.org/wiki/Universally_unique_identifier#Random_UUID_probability_of_duplicates

Leading me to believe that the same generated uuid was somehow assigned
to multiple rows.

I think I like Django's default primary key type better, but it was more
straightforward to add on a uuid type. It wasn't clear to me how to add
an AutoField to a model that already has a bunch of records, since
AutoField will throw an error when it doesn't have primary_key set
to True.