Fix user links in project report - PMT #105646
The user links on the project reports was giving a 404 for users whose
username isn't their uni. For newer users in PMT, the username is the
uni, but users who have been around for a while have a custom username.
I've fixed the template to look at the same attribute as the active
project report template, which doesn't have this problem.