merge tag functionality. PMT #92383
Handy for cleaning up tags. There are a bunch of very similarish tags
that people have put in and that makes tags less useful. Eg, people have
entered 'webservices', 'web-services', and 'web services' tags. Clearly,
those are supposed to be the same thing, but instead they are
separate. This makes it fairly simple to go into whichever are
considered non-canonical and merge them into the canonical tag.