parallel tests
Django 1.9 introduced the ability to run unit tests in parallel.

So I added a 'parallel-tests' target. On my desktop, with 8 cores, it
`make test` is 15 seconds while `make parallel-tests` is 2.5 seconds.

Leaving the existing on in place for now though as django-jenkins
doesn't yet know about parallel tests. So we'll need to either wait for
that or just replace django-jenkins and add our own coverage reporting.