handle missing target date on external add item
Addresses:
https://sentry.ccnmtl.columbia.edu/sentry-internal/dmt/group/1040/

If `target_date` is missing or empty, default to the milestone's target
date (same as regular add item).

This fixes the sentry error, but does not there's still the question of
how to handle an invalid target_date. I'm going to leave that as an open
issue for now.