Use auto_now for Item.last_mod - PMT #105540
To address this PMT, I've changed Item.last_mod to automatically update
whenever `Item.save()` is called. As a result, I've removed the `touch`
method and converted over the calls to `Item.touch()`.

Additionally, I've added a created_at field (which is allowed to be
null) for newly-created items. In the future, we could come up with a
data migration to tie each item's first event time to created_at, and
make it not-null.