fallback to User for comment
looking into this PMT #106983, I find this sentry error:
https://sentry.ccnmtl.columbia.edu/sentry-internal/dmt/group/1099/

it appears that recently some change is now setting the username on some
comments to the username of the `User`, not the `UserProfile` as it was
expecting, which in turn means that it now occasionally blows up when it
hits a user with a different username between the two.

I'm going to dig a bit further and see if I can figure out where the
change came in and fix it there, but in the meantime, this sets a
fallback to search by `User`.

Longer term, I also plan on just making it a proper foreign key,
probably straight to the `User`, but that will take a few steps, and
this fallback will make it simpler to implement the data migration that
will have to happen for that.