backfill Comment.author column
We're already double-writing, so after this, we can switch reads and
then delete the old username column.

There are 200k+ comments in the prod database, so this will take 5-10
minutes to run. On the upside, with Rolf out of the picture,
this *should* be fine. Jenkins will run it, it will take a while, but
nothing should time out.