Look for email in User.email as well as UserProfile.email
The user that caused this error has a User.email attribute but
no UserProfile.email attribute.

https://sentry.ccnmtl.columbia.edu/sentry-internal/dmt/group/1148/

This fix makes the logic more flexible, and there's many cases in PMT
where we look at either the User or UserProfile because one might be
empty. I realize this is kind of messy, but I think the cleanup work
belongs as a task on its own.