fix behave test
i think before it was "coincidentally" passing.

context.user.get_full_name() returns '' while the username is 'user_0'
or similar. On the page, 'user_0' is what gets displayed. I think
before, it was finding some other empty 'a' element and matching on
that, but removing the 'remove' link somehow changed that.