import models within tasks
not ideal, but necessary to avoid circular imports once we start wanting
to call tasks from models.py. From what I can tell, this is still the
recommended approach to dealing with the circular import:

https://groups.google.com/forum/#!topic/django-users/1mi7lGxWgvk

A better alternative might be using the `app` stuff and dynamically
loading models that way, but this gets us moving in the right direction
at least.