send email async with celery task
we shouldn't be sending email out as part of the request cycle. This
makes every mail sending operation go out async via celery.

This is also the first step toward dealing with:
https://pmt.ccnmtl.columbia.edu/item/108812/

It doesn't fix it, but now that it's a celery task, we will be able to
apply celery's built-in retry mechanism. (but I'll do that in a separate
PR later)