collapse verified items by default on kanban view
mature projects tend to accumulate hundreds of verified items in some
milestones and that just makes the page scroll a lot. *most* of the
time, it's only the open/in progress/resolved columns where you want to
see details, so this collapses the verified items by default to clean it
up visually a bit.