add project chat functionality
This essentially takes all of the Tala application and drops it into the
PMT.

This lets us have a little real-time websockets chat room for each
project in the PMT. That could be useful.

Mainly, since the only reason we have been keeping Tala around is as a
simple example of how to use the websockets stuff, this now replaces
that, and we can spin down Tala.

The UI, like in Tala, is still very basic. No presence, nothing
fancy. No that it's in the PMT and we might actually get people using
it, we might actually bother to implement that stuff.

prodoction/staging windsock settings are already salted out in
`local_settings.py`.