force default datetimes on some test messages
PMT #110068 seems to be because the database default/auto_now_add times
gets a timestamp and the python datetime defaults don't (or maybe
vice-versa, I'm not sure). So if the tests get run in the right window,
the message timestamp can appear to be in two different dates at once.

So, for the tests where that matters, just force the default to be set
from a python datestamp so it's consistent.