render chat messages with commonmark/emoji/etc.
commonmark setup was kind of a pain. Had to do the whole require.js
thing to make it work. And of course, that conflicted with the 'main.js'
require setup, so I put that in a template block so it could be
overridden selectively (none of that functionality needs to be loaded on
the chat page).

I attempted to get things working with require.js' multiversion support:
http://requirejs.org/docs/api.html#multiversion

but could not get it to load more than one of them no matter what I did.