eslint / jscs fixes
The eslint make target was defined in Makefile, overriding the target in
js.mk that correctly picks up all the JS directories. So I've uncovered
and fixed some eslint problems.

Also, it looks like jscs is being run on jenkins, not on travis.
Eventually I'll clean this up - we can just get rid of jshint/jscs in
favor of eslint.