Add 'internal' projects - PMT #111396
This adds the notion of 'internal' projects, I'm still not sure this is
the best name for this, I'm open to other suggestions. The goal is to
mark a few projects as "internal", and then make those display in
prominent places in the UI.

Internal projects are those that are specific to the CTL's operations,
like Facilities and Purchase Requests.