Updated URLs to match the Django > 2.0 functionalities
I have updated all the url() function to path() for non regular expression configured urls and to re_path() for the urls which are configured to match the regular expressions.
I have done this because while going through your project, I have seen in your requirements.txt that you are using Django > 2.0 and then I saw in urls, you are still using the functions that have been used before for Django < 2.0. So that's why I have updated the urls for better and cleaner code and functionalities.